Based on conversation with Joe Zack, changing the type of payload in process() to accept any so that raw XML or JSON posts are possible.  Process() now looks to see if payload is a struct and, if so, performs the same behavior as before.  If it's a simple value, it posts it as a body instead.
In addition, process() supports a new, third argument "headers" that is a struct of header names and values.  In REST APIs, often times you need to specify a Content-Type header for application/xml when passing XML.  It may also be necessary to pass an Accept or Authorization header and this argument allows gateway developers to pass in those details without having to modify the CFHTTP tag.